feat: socketio and apollo server combined
