+++ apps/L1G/tinder/tinder-backend/src/app/api/graphql/route.ts
@@ -1,5 +1,5 @@
 export const dynamic = 'force-dynamic';
 
-import { handler } from '../../../handler';
+import { handler } from '../../../server';
 
 export { handler as GET, handler as POST };
